+++ app/src/main/java/com/example/bodybuilder/data/DailyCalorieData/LossWeightData.kt
@@ -8,5 +8,5 @@ data class LossWeightData(
     @Json(name = "loss weight")
     var lossWeight: String?,
     @Json(name = "calory")
-    var calory: Int?
+    var calory: Float?
 )
