+++ app/src/main/java/com/example/bodybuilder/repository/Repository.kt
@@ -59,13 +59,13 @@ class Repository @Inject constructor(
             DailyCalorieData(
                 0.toFloat(),
                 DailyCalorieGoalsData(
-                    0,
-                    LossWeightData("",0),
-                    LossWeightData("",0),
-                    LossWeightData("",0),
-                    GainWeightData("", 0),
-                    GainWeightData("", 0),
-                    GainWeightData("", 0),
+                    0.toFloat(),
+                    LossWeightData("",0.toFloat()),
+                    LossWeightData("",0.toFloat()),
+                    LossWeightData("",0.toFloat()),
+                    GainWeightData("", 0.toFloat()),
+                    GainWeightData("", 0.toFloat()),
+                    GainWeightData("", 0.toFloat()),
                 )
             )
         )
